moves handlers into a handler file Relates #1
